Move `HttpStatusFilter` out of `Callbacks` module
The `HttpStatusFilter` class is more generally applicable than the
`Callbacks` module.

This commit moves it up a level in the module namespacing so that it's
referenced by `ActionClient::HttpStatusFilter`.
